requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9